pyproject.toml
[project]
name = "prefect-data-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "prefect>=3.4.1",
    "pymysql>=1.1.1",
    "python-dotenv>=1.1.0",
    "sqlalchemy>=2.0.41",
]
